241206(댓글 삭제기능 최종)

+++ blogproject_front-end/src/components/PostDetail.vue
@@ -18,31 +18,52 @@
         <p><strong>글쓴이:</strong> {{ post.name }}</p>
         <p>
           <strong>작성일:</strong>
-          {{ post.created_at ? formatDate(post.created_at) : "작성일 정보 없음" }}
+          {{
+            post.created_at ? formatDate(post.created_at) : "작성일 정보 없음"
+          }}
         </p>
         <p><strong>조회수:</strong> {{ post.views }}</p>
         <div v-html="post.content"></div>
         <!-- Quill에서 저장된 HTML 내용을 보여줌 -->
       </div>
       <div class="button-container">
-        <button @click="editPost" class="edit-btn">수정</button>
-        <button @click="deletePost" class="delete-btn">삭제</button>
+        <button v-if="post.email === currentUserId" @click="editPost" class="edit-btn">수정</button>
+        <button v-if="post.email === currentUserId" @click="deletePost" class="delete-btn">삭제</button>
       </div>
 
       <div class="comments-section">
-        <h3>댓글</h3>
+        <h3>댓글 {{ post.commentCount }}</h3>
         <form @submit.prevent="addComment">
-          <textarea v-model="newComment" placeholder="댓글을 작성하시오" required></textarea>
+          <textarea
+            v-model="newComment"
+            placeholder="댓글을 작성하시오"
+            required
+          ></textarea>
           <button type="submit">댓글 작성</button>
         </form>
 
         <div class="comments-list">
-          <div v-for="comment in comments" :key="comment.id" class="comment-item">
+          <div
+            v-for="comment in comments"
+            :key="comment.id"
+            class="comment-item"
+          >
             <p>{{ comment.content }}</p>
-            <small>작성자 : {{ comment.username }} | {{ comment.createdAt }}</small>
-            <button v-if="comment.userId === currentUserId" @click="deleteComment(comment.id)">삭제</button>
-          </div>
 
+            <div class="comment-footer">
+              <small
+                >작성자 : {{ comment.username }} |
+                {{ comment.createdAt }}</small
+              >
+              <!-- <button
+                v-if="comment.userId === currentUserId"
+                @click="deleteComment(comment.id)"
+              >
+                삭제
+              </button> -->
+              <img v-if="comment.userId === currentUserId" src="@/assets/delete-comment.jpg" alt="삭제" class="delete-icon" @click="deleteComment(comment.id)">
+            </div>
+          </div>
         </div>
       </div>
     </div>
@@ -56,8 +77,9 @@ export default {
   data() {
     return {
       comments: [],
-      newComment:"",
-      currentUserId: JSON.parse(localStorage.getItem("currentUser"))?.id || null,
+      newComment: "",
+      currentUserId:
+        JSON.parse(localStorage.getItem("currentUser"))?.email || null,
       post: {}, // 상세 글 데이터
       isLoggedIn: !!localStorage.getItem("currentUser"),
     };
@@ -66,7 +88,9 @@ export default {
   created() {
     this.getPostDetail();
     this.fetchComments();
+    console.log("현재 사용자 ID:", this.currentUserId); // 디버깅 코드
   },
+
   methods: {
     async getPostDetail() {
       const postId = parseInt(this.$route.params.id);
@@ -83,7 +107,6 @@ export default {
         this.post = response.data.post;
 
         await this.incrementViewCount(postId);
-
       } catch (error) {
         console.error(
           "게시물 불러오기 중 오류 발생:",
@@ -94,12 +117,16 @@ export default {
     },
     async fetchComments() {
       try {
-        const response = await axios.get(`http://localhost:3000/comments/${this.$route.params.id}`);
-        this.comments = response.data.map(comment => ({
-          ...this.comment,
+        const response = await axios.get(
+          `http://localhost:3000/comments/${this.$route.params.id}`
+        );
+        this.comments = response.data.map((comment) => ({
+          //...this.comment,
+          id: comment.id,
+          userId: comment.userId,
           username: comment.userId,
           createdAt: this.formatDate(comment.createdAt),
-          content: comment.content
+          content: comment.content,
         }));
       } catch (error) {
         console.error("댓글 불러오기 실패:", error);
@@ -109,36 +136,59 @@ export default {
     async addComment() {
       try {
         const token = localStorage.getItem("token");
-        const storedUser = JSON.parse(localStorage.getItem('currentUser'));
+        const storedUser = JSON.parse(localStorage.getItem("currentUser"));
         if (!storedUser || !storedUser.email) {
-          throw new Error('사용자 정보가 없습니다.');
+          throw new Error("사용자 정보가 없습니다.");
         }
 
         const response = await axios.post(
           `http://localhost:3000/comments/${this.$route.params.id}`,
-          { postId: this.$route.params.id, content: this.newComment, userEmail: storedUser.email },
+          {
+            postId: this.$route.params.id,
+            content: this.newComment,
+            userEmail: storedUser.email,
+          },
           {
             headers: { Authorization: `Bearer ${token}` },
           }
         );
-        this.comments.push(response.data);
-        this.newComment = "";
+        this.comments.push({
+          id: response.data.id,
+          userId: response.data.username,
+          username: response.data.username,
+          createdAt: this.formatDate(response.data.createdAt),
+          content: response.data.content,
+        });
+        this.newComment = '';
       } catch (error) {
         console.error("댓글 작성 실패:", error.response?.data || error.message);
-        alert('댓글 작성에 실패했습니다. 다시 시도해주세요.');
+        alert("댓글 작성에 실패했습니다. 다시 시도해주세요.");
       }
     },
 
     //댓글 삭제
     async deleteComment(commentId) {
+      alert('정말 이 댓글을 삭제하시겠습니까?');
       try {
-        const token = localStorage.getItem('token');
+        const token = localStorage.getItem("token");
+        const storedUser = JSON.parse(localStorage.getItem("currentUser"));
+        if (!storedUser || !storedUser.email) {
+          throw new Error("사용자 정보가 없습니다.");
+        }
+        const currentUserEmail = storedUser.email;
+
         await axios.delete(`http://localhost:3000/comments/${commentId}`, {
-          headers: { Authorization: `Bearer ${token}` },
+          headers: {
+            Authorization: `Bearer ${token}`,
+            "current-user": currentUserEmail,
+          },
         });
-        this.comments = this.comments.filter(comment => comment.id !== commentId);
-      } catch(error) {
-        console.log('댓글 삭제 실패:', error.response?.data || error.message);
+
+        this.comments = this.comments.filter(
+          (comment) => comment.id !== commentId
+        );
+      } catch (error) {
+        console.log("댓글 삭제 실패:", error.response?.data || error.message);
       }
     },
 
@@ -146,17 +196,20 @@ export default {
       try {
         await axios.put(`http://localhost:3000/detail/${postId}/views`);
       } catch (error) {
-        console.error('조회수 증가 에러 발생:',error.response ? error.response.data : error.message);
+        console.error(
+          "조회수 증가 에러 발생:",
+          error.response ? error.response.data : error.message
+        );
       }
     },
 
     // 글 수정 페이지로 이동
     editPost() {
-      const currentUser = localStorage.getItem('currentUser'); //현재 로그인한 사용자 가져오기
+      const currentUser = localStorage.getItem("currentUser"); //현재 로그인한 사용자 가져오기
 
       if (!currentUser) {
-        alert('로그인이 필요합니다.');
-        this.$router.push('/login');
+        alert("로그인이 필요합니다.");
+        this.$router.push("/login");
         return;
       }
       this.$router.push(`/edit/${this.post.id}`);
@@ -175,27 +228,29 @@ export default {
     async deletePost() {
       if (confirm("정말 이 글을 삭제하시겠습니까?")) {
         try {
-          const storedUser = JSON.parse(localStorage.getItem('currentUser'));
+          const storedUser = JSON.parse(localStorage.getItem("currentUser"));
           const currentUserEmail = storedUser.email;
           console.log(currentUserEmail);
-          console.log('storedUserEmail: ', storedUser.email);
-          // console.log('작성자 이메일: ', postAuthorEmail, '현재 사용자 이메일: ', currentUserEmail)
+          console.log("storedUserEmail: ", storedUser.email);
 
           await axios.delete(`http://localhost:3000/detail/${this.post.id}`, {
             headers: {
-              'current-user': currentUserEmail,
+              "current-user": currentUserEmail,
             },
           });
           alert("게시물이 삭제되었습니다.");
           this.$router.push("/list"); // 글 목록 페이지로 이동
         } catch (error) {
-          console.error("게시물 삭제 중 오류 발생:", error.response?.data || error.message);
+          console.error(
+            "게시물 삭제 중 오류 발생:",
+            error.response?.data || error.message
+          );
           alert("게시물 삭제에 실패했습니다.");
 
           if (error.response?.status === 403) {
-            alert('삭제 권한이 없습니다.');
+            alert("삭제 권한이 없습니다.");
           } else {
-            alert('게시물 삭제에 실패했습니다.');
+            alert("게시물 삭제에 실패했습니다.");
           }
         }
       }
@@ -209,7 +264,6 @@ export default {
       const hours = String(date.getHours()).padStart(2, "0");
       const minutes = String(date.getMinutes()).padStart(2, "0");
       return `${year}-${month}-${day} ${hours}:${minutes}`;
-      
     },
   },
   mounted() {
@@ -219,6 +273,16 @@ export default {
 </script>
 
 <style scoped>
+.delete-icon {
+  width: 15px;
+  height: 15px;
+  cursor: pointer;
+}
+
+.delete-icon {
+  opacity: 0.8;
+}
+
 .container {
   display: flex;
   justify-content: center;
@@ -332,10 +396,67 @@ div {
 }
 
 .comment-item {
-  margin-bottom: 10px;
-  padding: 5px;
+  padding: 10px 0;
   border-bottom: 1px solid #ddd;
 }
+
+.comment-item p {
+  margin: 0;
+  font-size: 16px;
+  line-height: 1.4;
+}
+
+.comment-footer {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin-top: 5px;
+  font-size: 14px;
+  color: #666;
+}
+
+/* 삭제 버튼 스타일 */
+.comment-footer .delete-btn {
+  background-color: #ff4d4f;
+  color: white;
+  border: none;
+  padding: 5px 10px;
+  border-radius: 5px;
+  cursor: pointer;
+  font-size: 14px;
+}
+
+.comment-footer .delete-btn:hover {
+  background-color: #d9363e;
+}
+
+/* .comment-item {
+  margin-bottom: 10px; 
+  padding: 5px;
+  border-bottom: 1px solid #ddd;
+
+  justify-content: space-between;
+  align-items: center;
+  display: flex;
+}
+
+.comment-item small {
+  margin-right: auto;
+  color: #666
+}
+
+.comment-item button {
+  background-color: #ff4d4f;
+  color: white;
+  border: none;
+  padding: 5px 10px;
+  border-radius: 5px;
+  cursor: pointer;
+}
+
+.comment-item button:hover {
+  background-color: #d9363e;
+} */
 
 textarea {
   width: 100%;
